Update README.md
The aim of this Embedded system is to "Design an intelligent billboard that displays useful advertisements in relation to the people who are waiting to cross by the traffic lights"  Consider, a traffic light with only two colors. YELLOW for  (cars are moving) and RED for (cars are not moving / Cars are waiting),
"YELLOW" indicates that the pedestrian or biker is approaching a road crossing at the time the traffic signal is "YELLOW" (Cars are moving). 

The time is shown on a billboard on the side of the building. The pedestrian or cyclist presses a button to indicate that he or she want to cross the street. From that point on, the traffic light begins to blink (YELLOW BLINKING) every 30 seconds for the next 30 seconds. During that time frame, the billboard will display an advertisement targeted specifically at pedestrians and cyclists.

When the traffic light becomes RED after 30 seconds, cars come to a stop and wait for the signal to revert back to YELLOW. For approximately 1 minute, pedestrians and cyclists are permitted to cross the street. During that particular period, the billboard will display an advertisement that is targeted specifically at car drivers.
When pedestrians and cyclists are permitted to cross, a sound should be played to alert disabled people that they are permitted to cross.